refactor(bets): extract main event bet types
